Desktop/IT/muctr/muctr_backend/muctr_back/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-better-admin-arrayfield==1.4.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-postgres-extensions==0.9.3
djangorestframework==3.14.0
jsonfield==3.1.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4